Starting to work

+++ PA4/txtHelper/char2OneHot.m
@@ -2,6 +2,6 @@ function [ one_hot ] = char2OneHot( input_char )
 %CHAR2ONEHOT Part1 used to convert training date into 1hot
 
 one_hot = zeros(256,1);
-one_hot(uint(input_char)) = 1;
+one_hot(uint16(input_char)) = 1;
 end
